Производит рефакторинг кода
