+++ src/mock/sort.js
@@ -3,7 +3,7 @@ import { DAY_FORMAT, TIME_FORMAT } from '../utils/utils';
 
 const SortType = {
   DAY: 'day',
-  EVENTS: 'event',
+  EVENT: 'event',
   TIME: 'time',
   PRICE: 'price',
   OFFERS: 'offer'
@@ -11,7 +11,7 @@ const SortType = {
 
 const sort = {
   [SortType.DAY]: (points) => points.sort((a, b) => dayjs(b.dateFrom).format(DAY_FORMAT) - dayjs(a.dateFrom).format(DAY_FORMAT)),
-  [SortType.EVENTS]: (points) => points,
+  [SortType.EVENT]: (points) => points,
   [SortType.TIME]: (points) => points.sort((a, b) => dayjs(b.dateFrom).format(TIME_FORMAT) - dayjs(a.dateFrom).format(TIME_FORMAT)),
   [SortType.PRICE]: (points) => points.sort((a, b) => b.basePrice - a.basePrice),
   [SortType.OFFERS]: (points) => points,
@@ -21,5 +21,3 @@ const generateSort = (points) => Object.entries(sort).map(([sortName, sortPoint]
   name: sortName,
   points: sortPoint(points),
 }));
-
-export { generateSort};
