+++ src/utils/const.js
@@ -1,0 +1,10 @@
+const SORT_TYPE = ['day', 'event', 'time', 'price', 'offer'];
+
+const POINT_PRICE = {
+  min: 100,
+  max: 2000,
+};
+
+const TYPES_OF_TRANSPORT = ['taxi', 'bus', 'train', 'ship', 'drive', 'flight', 'check-in', 'sightseeing', 'restaurant'];
+
+export { SORT_TYPE, POINT_PRICE, TYPES_OF_TRANSPORT };
