+++ src/utils/date.js
@@ -1,0 +1,9 @@
+const DAY_FORMAT = 'DD';
+const DATE_FORMAT = 'MMM DD';
+const TIME_FORMAT = 'HH:mm';
+const DATE_TIME_FORMAT = 'YYYY-MM-DDTHH:mm';
+const DATE_TIME_INPUT_FORMAT = 'DD/MM/YY HH:mm';
+
+const PRESENT_DATE = new Date();
+
+export {DAY_FORMAT, DATE_FORMAT, TIME_FORMAT, DATE_TIME_FORMAT, DATE_TIME_INPUT_FORMAT, PRESENT_DATE};
