+++ src/view/header-views/filters-view.js
@@ -4,9 +4,9 @@ const createItemFilterTemplate = (filterItem, isChecked) => {
   const {name, count} = filterItem;
   return (
     `<div class="trip-filters__filter">
-      <input id="filter-${name}" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything"
+      <input id="filter-${name}" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="${name}"
         ${isChecked ? 'checked' : ''}${count === 0 ? 'disabled' : ''}>
-      <label class="trip-filters__filter-label" for="filter-everything">${name}</label>
+      <label class="trip-filters__filter-label" for="filter-${name}">${name}</label>
     </div>`
   );
 };
