+++ src/view/main-views/edit-point-view.js
@@ -1,5 +1,6 @@
 import AbstractView from '../../framework/view/abstract-view';
-import { humanizeDate, DATE_TIME_INPUT_FORMAT } from '../../utils/utils';
+import { DATE_TIME_INPUT_FORMAT } from '../../utils/date';
+import { humanizeDate } from '../../utils/utils';
 
 function createItemEditPointTemplate (point, pointDestination, pointOffers) {
   const {type, basePrice, dateFrom, dateTo} = point;
