+++ src/view/main-views/sort-view.js
@@ -1,19 +1,14 @@
 import AbstractView from '../../framework/view/abstract-view';
 
-const createSortItemTemplate = (sortItem, isChecked) => {
-  const {name} = sortItem;
-
-  return (
-    `<div class="trip-sort__item  trip-sort__item--${name}">
-      <input id="sort-${name}" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-${name}" ${isChecked ? 'checked' : ''}>
-      <label class="trip-sort__btn" for="sort-${name}">${name}</label>
-    </div>`
-  );
-};
+const createSortItemTemplate = (sortName, isChecked) =>
+  `<div class="trip-sort__item  trip-sort__item--${sortName}">
+    <input id="sort-${sortName}" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-${sortName}" ${isChecked ? 'checked' : ''}>
+    <label class="trip-sort__btn" for="sort-${sortName}">${sortName}</label>
+  </div>`;
 
 const createEventsSortTemplate = (sorting) =>
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
-    ${sorting.reduce((prev, sortItem) => `${prev}${createSortItemTemplate(sortItem)}`, '')}
+    ${sorting.reduce((prev, sortName) => `${prev}${createSortItemTemplate(sortName)}`, '')}
   </form>`;
 
 export default class EventsSortView extends AbstractView {
@@ -22,7 +17,6 @@ export default class EventsSortView extends AbstractView {
   constructor (sorteredPoints) {
     super();
     this.#sorting = sorteredPoints;
-
   }
 
   get template() {
